Load 3.3.699 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@1473 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/slaballoc.c
@@ -475,7 +475,7 @@ struct slab_s
                              * TODO:: slabentry_t.
                              */
     mslab_t * prev, * next;  /* Link pointers for slab list. */
-    unsigned short numAllocated;  /* Number of blocks allocated. */
+    unsigned long numAllocated;  /* Number of blocks allocated. */
     word_t * freeList;      /* List of free blocks. */
     word_t   blocks[1];     /* First word of the allocatable memory. */
 };
